move FormatDependent outside of helpers
