+++ app/helpers/custom_fields_helper.rb
@@ -27,44 +27,6 @@
 #++
 
 module CustomFieldsHelper
-  class FormatDependent
-    CONFIG = {
-      allowNonOpenVersions: [:only, %w[version]],
-      defaultBool: [:only, %w[bool]],
-      defaultLongText: [:only, %w[text]],
-      defaultText: [:except, %w[list bool date text user version]],
-      length: [:except, %w[list bool date user version link]],
-      multiSelect: [:only, %w[list user version]],
-      possibleValues: [:only, %w[list]],
-      regexp: [:except, %w[list bool date user version]],
-      searchable: [:except, %w[bool date float int user version]],
-      textOrientation: [:only, %w[text]]
-    }.freeze
-
-    def self.stimulus_config
-      CONFIG.map { |target_name, (operator, formats)| [target_name, operator, formats] }.to_json
-    end
-
-    attr_reader :format
-
-    def initialize(format)
-      @format = format # rubocop:disable Rails/HelperInstanceVariable
-    end
-
-    def attributes(target_name)
-      operator, formats = CONFIG[target_name.to_sym]
-
-      fail ArgumentError, "Unknown target name #{target_name}" unless formats
-
-      visible = operator == :only ? format.in?(formats) : !format.in?(formats)
-
-      ApplicationController.helpers.tag.attributes(
-        "data-admin--custom-fields-target": target_name,
-        hidden: !visible
-      )
-    end
-  end
-
   def custom_fields_tabs
     [
       {
